Console output language is converted to English
