+++ AES_Robot_1/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Ahmet Emin SA\u011eLIK\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=C:\\Users\\ahmet\\AppData\\Roaming\\NetBeans\\21\\build.properties
