+++ AES_Robot_1/nbproject/private/private.xml
@@ -2,6 +2,8 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/D:/projects/Apache-NetBeanse/PathFinder-1/AES_Robot_1/src/aes_robot1/AES_Robot_1_English.java</file>
+        </group>
     </open-files>
 </project-private>
