+++ AES_Robot_1/robot.txt
@@ -1,0 +1,26 @@
+------------------------------------------------
+----------------  LABİRENTİMİZ  ----------------
+------------------------------------------------
+|☺| | | | | | | 
+|X| |X| |X| | | 
+| | | | | | | | 
+| | |X| |X| |X| 
+| | | | |☯| | | 
+
+------------------------------------------------
+-----------  LABİRENTİMİZ  BİTMİŞTİR -----------
+------------------------------------------------
+
+Arabamiz 1 br saga ilerledi [0,3]
+!!!!!Maximum sag tarafa ilerledi, Sola Gitmek icin referans noktasina donuluyor
+Arabamiz  1birim sola kaydı ve referans noktasina geri döndü  
+Arabamiz 1 br sola ilerledi [0,1]Arabamiz 1 br sola ilerledi [0,0]Arabamiz  2 br yukari ilerledi
+!!! Referansa Dönüldü --- Konumu : [2,0]--> İlerlenecek yon SAG taraf
+Arabamiz 1 br saga ilerledi [2,1]
+Arabamiz 1 br saga ilerledi [2,2]
+Arabamiz 1 br saga ilerledi [2,3]
+Arabamiz  2 br yukari ilerledi
+!!! Referansa Dönüldü --- Konumu : [4,3]--> İlerlenecek yon SAG taraf
+Referans noktamız en sagda doğruca sol tarafa doğru ilerlemeye başlıyoruz!!!!!!!!!!!!!!!!!!!!!!!!!Arabamiz 1 br sola ilerledi [4,2]Arabamiz 1 br sola ilerledi [4,1]Arabamiz 1 br sola ilerledi [4,0]Arabamiz cıkısa ulasmistirSola ilerlenen toplam Birim : 6
+Saga ilerlenen toplam Birim : 4
+
